requirements.txt
numpy==1.21.1
Pillow==8.3.1
protobuf==3.17.3
six==1.16.0
